feat: 예약이 존재하는 시간 삭제 시 예외처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,18 +2,22 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
+import roomescape.exception.ReservationTimeInUseException;
 
 @Service
 public class ReservationTimeService {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<ReservationTimeResponseDto> findAll() {
@@ -28,6 +32,9 @@ public class ReservationTimeService {
     }
 
     public boolean delete(Long id) {
+        if (reservationDao.existsByTimeId(id)) {
+            throw new ReservationTimeInUseException(id);
+        }
         return reservationTimeDao.deleteById(id) > 0;
     }
 }
